+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 2.30 branch
 ### 2.30.11
 * the `topbar` element has been made permanently compatible with the new bootstrap;
+* the `default` layout will correctly use the `topbar.php` and `footer.php` elements if they are present in the app,
+  otherwise it will use the plugin ones;
 * the `default` layout can automatically establish whether to use a sidebar and whether to directly use the 
   css files (`layout.css` and `contents.css`) from the application webroot;
 * the `login` layout first checks if the `logo_login.png` file exists, otherwise it uses the default logo;